Added files via upload
Everything that I wanted to do was edit the actual Bootstrap Carousel file itself. This was very challenging considering that that's not how you're supposed to work with this set of files in particular. So I just changed some text and reordered the nav. I feel satisfied and I'm ready to lean more. :D